Move DirectoryMoc to make_mocs.py
