+++ .github/scripts/update-mocs.py
@@ -2,9 +2,7 @@ import os
 import sys
 import argparse
 
-from make_mocs import filter_directories, moc_file_path_for_directory, moc_base_name_for_directory, \
-    make_moc_for_directory_with_delimiters, update_existing_moc
-from utils import get_template
+from make_mocs import DirectoryMoc, filter_directories
 
 
 class VaultMoc:
@@ -23,36 +21,6 @@ class VaultMoc:
             dirs.sort()
             directory_moc = DirectoryMoc(root, dirs, files)
             directory_moc.generate_moc()
-
-
-class DirectoryMoc:
-    """Class to manage the MOC for a single directory"""
-
-    def __init__(self, root, dirs, files):
-        self.root = root
-        self.moc_file_path = moc_file_path_for_directory(root)
-        self.dirs = dirs
-        self.files = files
-
-    def generate_moc(self):
-        new_moc_content_with_delimiters = make_moc_for_directory_with_delimiters(self.root, self.dirs, self.files)
-        if os.path.exists(self.moc_file_path):
-            self.rewrite_existing_moc_file(new_moc_content_with_delimiters)
-        else:
-            self.write_new_moc_file(new_moc_content_with_delimiters)
-
-    def rewrite_existing_moc_file(self, new_moc_content_with_delimiters):
-        with open(self.moc_file_path, 'r') as input:
-            initial_content = input.readlines()
-        with open(self.moc_file_path, 'w') as output:
-            output.write(update_existing_moc(initial_content, new_moc_content_with_delimiters))
-
-    def write_new_moc_file(self, new_moc_content_with_delimiters):
-        template = get_template("directory_moc")
-        moc_base_name = moc_base_name_for_directory(self.root)
-        new_content = template.render(title=moc_base_name, list_of_files_and_dirs=new_moc_content_with_delimiters)
-        with open(self.moc_file_path, 'w') as output:
-            output.write(new_content)
 
 
 def main(argv=sys.argv[1:]):
